tests/wpt/web-platform-tests/css/css-writing-modes/parsing/direction-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Writing Modes Level 3: parsing direction with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#propdef-direction">
<meta name="assert" content="direction supports only the grammar 'ltr | rtl'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("direction", "auto");
test_invalid_value("direction", "ltr rtl");
</script>
</body>
</html>
